x11/plasma5-kactivitymanagerd: Add missing USE_QT=sql-sqlite3:run
In a full KDE desktop qt5-sqldrivers-sqlite3 normally ends up being
pulled in via a long chain of dependencies to qt5-help, which has it as
an explicit dependency. However, in more minimal environments, we can
end up with the package being missing, which causes kactivitymanagerd to
immediately exit with a fatal error from being unable to open the
database. If it exits too quickly this can even cause plasmashell to
abort its startup, otherwise it leads to repeated kactivitymanagerd
crashes that, by default, get reported by Dr Konqi as a flood of popups.