Use LOCALBASE64 not LOCALBASE to set the default PYTHONBASE.
In a poudriere jail the script ports_env.sh sets the environmental varaible PYTHONBASE.
In a CheriABI jail this should be set to the LOCALBASE64 (as a CheriABI Python package is currently not available).
Once the value of PYTHONBASE is set in /etc/make.conf.ports_env it is never updated to the correct value in python.mk.