x11-wm/plasma5-kwin: Add a libseat-based session backend
ConsoleKit2 requires GLib, which isn't ported, and logind is specific to
Linux, so we're using seatd instead and need a backend that can talk to
that. Technically libseat can also talk to the others so those specific
backends are a bit redundant, but they do provide richer functionality
as libseat abstracts them to a simpler common interface.

This allows unprivileged users to run KWin without needing input devices
to be accessible by them, as the seat manager (here seatd) opens them on
its behalf.