devel/dbus: Apply patch from upstream for incorrect va_arg use
Without optimisation the code ends up loading one past the end of the
variadic argument array and taking a bounds fault on Morello. With
optimisation the load ends up being sunk such that it's converted to the
corrected code where the load is guarded by the previous value not being
NULL, since that's the only case in which the loaded value is used.