CHERI: Detect and replace gnulib's broken stdin.h
gnulib insists on installing a stdin.h wrapper that among other
things) redefines (u)intptr_t to (unsigned) long with a hard to mix
of typedefs and #defines.  In practice our stdint.h is fine and does
not include these bugs.

Because gnulib bits are embedded in various software distributions
(gettext alone includes FOUR copies), unconditionally search for
and replace instances of the stdint.in.h templace.